Updated FinalProject.cpp
Added some check variables to see if files open, and use these checks to modify the menu. If the message file does not open, interacting with the menu will seg fault. (Aside from trivia, but it's hard to do trivia without being able to print anything). Without the cpi file, printing by highest gross adjusted for inflation seg faults. If question file does not open, movie trivia seg faults. Added if statements so that the user can still print lots of info about the movies with just the movie file without having to worry about seg faults. This is really only a problem if the user doesn't run the file correctly, but it might still be nice to have. Not completely necessary though.